Fix merge conflict
